Adds comments, changes to sorting process, accommodates Levenshtein's edit distance algorithim
Changes made to accommodate Levenshtein's edit distance code that will act as the pivotal step to my sorting system, improving the order of relevance - to be added shortly by my partner, collaborator and contributor Ali Abdullah. I have also improved upon the existing sorting system which will act as a major component in improving accuracy of the engine. It will rely heavily on the distance between  characters in the objects in the list input as observed in the very same characters in the search input in the right order. Thus eliminating discrepancies in results when successful close searches that lie further down the string are to be run. Finally comments are now added at the requests by fellow contributors.